improving codebase, splitting into files
